+++ README.md
@@ -1,7 +1,5 @@
 # To-Do-List
 
-![screenshot](images/screenshot.png)
-
 This is my To-Do List project in the Microverse curriculum.
 
 ## Built With
