+++ src/style.css
@@ -113,7 +113,6 @@ section .task-section {
 
 .hide2 {
   display: flex;
-
 }
 
 .task-item {
